tools/deletebyquery/requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cassandra-driver==3.9.0
decorator==4.0.11
enum34==1.1.6
futures==3.1.1
ipython==5.3.0
ipython-genutils==0.2.0
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
requests==2.13.0
scandir==1.5
semver==2.7.6
simplegeneric==0.8.1
six==1.10.0
solrcloudpy==2.4.1
traitlets==4.3.2
wcwidth==0.1.7
